+++ client/src/App.jsx
@@ -3,7 +3,7 @@ import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
 import { BrowserRouter as Router } from 'react-router-dom'; 
-import Footer from './Components/Footer/Footer'; 
+import Footer from './Pages/Components/Footer'; 
 
 function App() {
   const [count, setCount] = useState(0)
