+++ client/src/Pages/Components/Modules/Images/CUT-logo.png
